+++ astropy/io/ascii/core.py
@@ -869,7 +869,7 @@ def convert_numpy(numpy_type):
     else:
         converter_type = AllType
 
-    def bool_converter(vals):
+    def bool_converter(vals, fast):
         """
         Convert values "False" and "True" to bools.  Raise an exception
         for any other string values.
@@ -889,8 +889,11 @@ def convert_numpy(numpy_type):
             raise ValueError('bool input strings must be only False or True')
         return trues
 
-    def generic_converter(vals):
-        return numpy.array(vals, numpy_type)
+    def generic_converter(vals, fast):
+    	if fast:
+    		return vals.astype(numpy_type, copy = False)
+    	else:
+        	return numpy.array(vals, numpy_type)
 
     converter = bool_converter if converter_type is BoolType else generic_converter
 
@@ -929,7 +932,6 @@ class BaseOutputter(object):
             # to set the defaults, otherwise use the generic defaults.
             default_converters = ([convert_numpy(col.dtype)] if col.dtype
                                   else self.default_converters)
-
             # If the user supplied a specific convert then that takes precedence over defaults
             converters = self.converters.get(col.name, default_converters)
 
@@ -946,7 +948,7 @@ class BaseOutputter(object):
                     converter_func, converter_type = col.converters[0]
                     if not issubclass(converter_type, col.type):
                         raise TypeError('converter type does not match column type')
-                    col.data = converter_func(col.str_vals)
+                    col.data = converter_func(col.str_vals, False)
                     col.type = converter_type
                 except (TypeError, ValueError) as err:
                     col.converters.pop(0)